seeders users&productImagenes

+++ src/database/migrations/20230514032335-create_products_table.js
@@ -41,12 +41,12 @@ module.exports = {
         type: Sequelize.STRING(20),
       },
     purchasePrice: {
-        type: Sequelize.DECIMAL(11, 10),
+        type: Sequelize.INTEGER,
               allowNull: false
         },
         
     salePrice: {
-        type: Sequelize.DECIMAL(11, 10),
+        type: Sequelize.INTEGER,
               allowNull: false
         },
